[#18] Rename BackgroundImage to MainView
